Functional dynamic cropping with front camera
Fix the jitter in front-face dynamic cropping by
changing from float to int. The use of floating point
would corrupt the crop operation.

Also add protection where we change the lower
bound of the crop box when the ankles are not
found, so that if they appear later they’re more likely
to be detected.

+++ ViroRenderer/VROBodyTrackeriOS.h
@@ -192,8 +192,8 @@ private:
      Perform a crop and pad of the given image using the dynamic crop box,
      and return the result.
      */
-    CVPixelBufferRef performCropAndPad(CVPixelBufferRef image, float *outCropX, float *outCropY,
-                                       float *outCropWidth, float *outCropHeight);
+    CVPixelBufferRef performCropAndPad(CVPixelBufferRef image, int *outCropX, int *outCropY,
+                                       int *outCropWidth, int *outCropHeight);
     
     /*
      Derive the full body bounds from the given pose.
